Refactored merge from mongo and sql branch
