+++ views/editMovie.pug
@@ -34,4 +34,4 @@ html(lang="en")
                 p Imagen:
                     input(type="text", value=`${movie.imagen}` name="imagen")
                 input(type="submit" value="Guardar")
-        script(src='./js/main.js')
+        script(src='./js/edit_movie.js')
